+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 - **Automap position and scale being reset when changing modes**
 - **Crosshair not being drawn if the chasecam mode were set** even if the chasecam itself were disabled
 - **Zoom being forcefully disabled if any player were dead**
+- **HUD key not updating NUGHUD view offset**
 - **Explosion shake and FOV effects not being reset when restarting levels on death**
 - **_Selective Fuzz Darkening_ not being disabled in Strict Mode**
 - **Changed save version to `Nugget 3.0.0`** (was `Nugget 2.4.0` by mistake)